Move @id validations into module
